requirements.txt
ansible==2.9.6
apache-libcloud==2.8.0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.8.1
attrs==19.3.0
bcrypt==3.1.7
blinker==1.7.0
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==2.2.12
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==3.0.2
future==0.18.2
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
httplib2==0.14.0
idna==2.8
importlib-metadata==7.0.1
iotop==0.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.9.4
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
netaddr==0.7.19
netifaces==0.10.4
ntlm-auth==1.1.0
numpy==1.23.5
oauthlib==3.1.0
olefile==0.46
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.6.0.66
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==4.25.3
psutil==5.5.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==3.0.1
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
